# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marasem API",
        "description": "This API supports social media login with Google, Facebook, and Behance. Users are redirected to the provider's login page and upon successful authentication, they are logged into the app and issued a token.",
        "contact": {
            "email": "amrfayez.247@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server for API routes"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server for Web routes"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/add-address": {
            "post": {
                "tags": [
                    "Address"
                ],
                "summary": "Add a new address for the authenticated user",
                "operationId": "6b5169034a08e2084b2ad42035275e1c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "city",
                                    "zone",
                                    "address"
                                ],
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Cairo"
                                    },
                                    "zone": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Downtown"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "123 Main St"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John's Office",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "+1234567890",
                                        "nullable": true
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "maxLength": 5,
                                        "example": "+20",
                                        "nullable": true
                                    },
                                    "is_default": {
                                        "description": "Set as default address",
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "Cairo"
                                        },
                                        "zone": {
                                            "type": "string",
                                            "example": "Downtown"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "123 Main St"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John's Office",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "+1234567890",
                                            "nullable": true
                                        },
                                        "country_code": {
                                            "type": "string",
                                            "example": "+20",
                                            "nullable": true
                                        },
                                        "is_default": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/addresses/{id}": {
            "put": {
                "tags": [
                    "Address"
                ],
                "summary": "Update an existing address",
                "operationId": "46a5db4761fba5463c3510212d68b8ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Giza",
                                        "nullable": true
                                    },
                                    "zone": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Dokki",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "456 Elm St",
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John's Apartment",
                                        "nullable": true
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "+201234567890",
                                        "nullable": true
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "maxLength": 5,
                                        "example": "+20",
                                        "nullable": true
                                    },
                                    "is_default": {
                                        "description": "Set as default address",
                                        "type": "boolean",
                                        "example": true,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Address updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Address updated successfully."
                                        },
                                        "address": {
                                            "$ref": "#/components/schemas/Address"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Address"
                ],
                "summary": "Delete an address",
                "operationId": "7d0377e79c03534675ea7ce1b91c86e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the address to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Address deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Address not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Address not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Address": {
                "required": [
                    "user_id",
                    "city",
                    "zone",
                    "address"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the address",
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "The ID of the user associated with the address",
                        "type": "integer",
                        "example": 5
                    },
                    "city": {
                        "description": "The city of the address",
                        "type": "string",
                        "example": "Cairo"
                    },
                    "zone": {
                        "description": "The zone of the address",
                        "type": "string",
                        "example": "Downtown"
                    },
                    "address": {
                        "description": "The detailed address",
                        "type": "string",
                        "example": "123 Main Street"
                    },
                    "phone": {
                        "description": "Phone number for the address",
                        "type": "string",
                        "example": "1234567890"
                    },
                    "country_code": {
                        "description": "Country code for the address's phone",
                        "type": "string",
                        "example": "+20"
                    },
                    "name": {
                        "description": "Custom name for the address",
                        "type": "string",
                        "example": "John's Office"
                    },
                    "is_default": {
                        "description": "Indicates if this is the default address",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your Bearer token to access secured endpoints.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}